GH-39808: [C++][Parquet] Test out-of-order pre-buffer eviction
Move ReadCacheEvictionState to reader_internal.h and split its prefix-advance logic into a reader-free MarkDecodedAndGetEvictOffset() method that returns the evict-before offset (or nullopt). RowGroupDecoded is now a thin wrapper, so production behavior is unchanged while the out-of-order watermark advance is unit-testable without a reader.

Add a deterministic ReadCacheEvictionStateOutOfOrder test driving that method directly, covering out-of-order completion, prefix jumps, and idempotent re-decode.

Also drop now-dead includes (<mutex> in reader.cc once the class moved out; <limits>/<map>/<mutex>/<set> in file_reader.cc, left over from the removed run-merging members).
