+++ cpp/src/parquet/arrow/reader_internal.h
@@ -22,6 +22,8 @@
 #include <deque>
 #include <functional>
 #include <memory>
+#include <mutex>
+#include <optional>
 #include <unordered_set>
 #include <utility>
 #include <vector>
@@ -132,6 +134,60 @@ Status TransferColumnData(::parquet::internal::RecordReader* reader,
                           const std::shared_ptr<::arrow::Field>& value_field,
                           const ColumnDescriptor* descr, const ReaderContext* ctx,
                           std::shared_ptr<::arrow::ChunkedArray>* out);
+
+// ----------------------------------------------------------------------
+// Pre-buffer eviction
+
+// GH-39808: releases pre-buffered cache entries as row groups are decoded.
+// RowGroupDecoded() is called (possibly out of order, from readahead
+// continuations) once a row group's batches are produced. It advances a
+// watermark over the contiguous prefix of completed row groups and evicts every
+// cache entry ending before the lowest byte any not-yet-completed row group
+// still needs. A coalesced entry spanning a row-group boundary is freed once
+// the watermark passes its end.
+class ReadCacheEvictionState {
+ public:
+  // evict_before_offsets[i] = lowest byte offset row groups i..n-1 (in
+  // generator order) still need; evict_before_offsets[n] == INT64_MAX.
+  // Must be non-empty: it always holds the n row-group offsets plus the
+  // INT64_MAX sentinel.
+  explicit ReadCacheEvictionState(std::vector<int64_t> evict_before_offsets)
+      : evict_before_offsets_(std::move(evict_before_offsets)),
+        completed_(evict_before_offsets_.size() - 1, false) {}
+
+  void RowGroupDecoded(ParquetFileReader* reader, size_t row_group_index) {
+    if (auto evict_before = MarkDecodedAndGetEvictOffset(row_group_index)) {
+      reader->EvictPreBufferedDataBefore(*evict_before);
+    }
+  }
+
+  // Marks row_group_index decoded and advances the contiguous completed prefix.
+  // Returns the offset to evict before when the prefix advanced, or std::nullopt
+  // when nothing newly became evictable (index out of range, already completed,
+  // or the prefix did not grow). Separated from RowGroupDecoded so the
+  // out-of-order watermark logic is unit-testable without a reader.
+  std::optional<int64_t> MarkDecodedAndGetEvictOffset(size_t row_group_index) {
+    std::lock_guard<std::mutex> lock(mutex_);
+    if (row_group_index >= completed_.size() || completed_[row_group_index]) {
+      return std::nullopt;
+    }
+    completed_[row_group_index] = true;
+    const size_t old_prefix = completed_prefix_;
+    while (completed_prefix_ < completed_.size() && completed_[completed_prefix_]) {
+      ++completed_prefix_;
+    }
+    if (completed_prefix_ == old_prefix) {
+      return std::nullopt;
+    }
+    return evict_before_offsets_[completed_prefix_];
+  }
+
+ private:
+  std::mutex mutex_;
+  std::vector<int64_t> evict_before_offsets_;
+  std::vector<bool> completed_;
+  size_t completed_prefix_ = 0;
+};
 
 }  // namespace arrow
 }  // namespace parquet
