+++ cpp/src/parquet/file_reader.cc
@@ -20,12 +20,8 @@
 #include <algorithm>
 #include <cstdint>
 #include <cstring>
-#include <limits>
-#include <map>
 #include <memory>
-#include <mutex>
 #include <ostream>
-#include <set>
 #include <string>
 #include <unordered_map>
 #include <utility>
